requirements.txt
python-dotenv
telebot
openai==0.27.0